fix: Remove --force-overwrite option
The --force-overwrite option has proven to cause more problems than the
ones it solves, as it can damage content and create irrecoverable
situations.

In the situations where something goes very wrong and it's not possible
to resume a sync, it's better to create a new space, resync from the
source space and continue from there.

BREAKING CHANGE: removed --force-overwrite option